Modified getIncludeParams to return "" when paths is empty
If no include paths are specified in the configuration, getIncludeParams would return "-I" which would be included in the list of gfortran arguments. This appeared directly before the file path which caused gfortran to complain that no input files were specified. This change allows the include path specification to be removed if there are no include paths specified.