Replaces deprecated debugger interface.
As stated in issue #9037 in github.com/microsoft/vscode
is the way to go. Fortran does not do anything different from
the inhouse C++ debugger, hence we should only have the
extension allow breaking points and nothing else.
This of course means that we will be depending on
Microsoft's C++ extension for the debugger.