Fix overzealous OpenMP regex.
The OpenMP regex did not allow for the OpenACC syntax scope to trigger.
This has now been fixed and both should be displayed correctly.

The only outstanding problem is that the unittest (.snap) does not
trigger the right scopes, which means open acc/openmp are not
tested thoroughly. I look into it