Removes paths.ts which checks if binary is in PATH
This has now been replaced with `which` which is also platform agnostic.
The only difference is that `which` does not cache the result of the
query. However, realistically speaking calling which once per new
file being opened is a trivial cost.
Moreover, already opened files are cached automatically so no `which` calls
are made when changing between opened files.