Changed to eslint
Added back the husky and lint-staged although we don't really use them.

The name of Function had to be changed to FortranFunction since it clashed
with the TypeScript Function.
The `parseMe` var in the interface was deleted since it was never
used and it was causing errors due to the SymbolParser interface being empty.