Formats all files in the repo
The lint-stage hook sould now be able to work when we commit.
If a commit fails then run `npm run lint-fix` to try and autofix
and then `npm run format`.

All of these actions should be run
automatically via the git hook but I am listing the steps in case
the hook fails.