Fixes error with fast-glob and lack of dir access
When the read permissions are missing from a directory, `fast-glob`
throws a EACCES error and if you try and suppress it, it ends up with
a race condition. This is most likely a bug and I have opened an issue.

In the meantime, we are using `glob` which is noticeably slower to
recover.