Fixes formatter detection check
Also adds output on successfull install.
There is still a problem with the install tool returning a promise
that is being fullfied after we have executed the formatter
(that is if the tool is missing) which results into the user having
to call the formatter twice to format the document.