Removes usage of yarn
We can still use yarn if we want but the lockfile will not be stored anymore.
The reason being that yarn does not work with the git hooks for
precommits we have setup. We could probably get it to work
but npm works already so one package manages is enough.