fix: include `fypp` rather than compiler paths
The linter and `fypp` include paths are quite different and should
be set using different options.

Direct comment from PR:
"""
The include paths for fypp (where fypp include files should be looked for) are typically pointing to folders in a projects source tree. While the -I options, you pass to a compiler in order to look up mod-files, are typically pointing to folders in the projects build tree. So, strictly speaking, there are completely different and should be treated differently. On the other hand, passing -I options meant for the compiler to fypp won't make any harm, as long as they do not contain source files included via #:include in any fypp-processed files. Chances for that are pretty low.
"""