feat: makes path caching& glob resolution reusable
We change the way caching is done. A class was created
that holds the globs and the resolved and expanded paths.

On the linter side, a `Map` (has `O(1)` access) is used to
store all the VS Code options that would need a cache e.g.
`cache.set('fortran.linter.option', new GlobPaths(globs))`.

Possible performance improvements:
- Use Sets instead of Arrays
- Prefetch certain directories

TODO:
- Add unittests for linter integration

No changes to the API have been done