feat: try and discover a valid Python interpreter
This is particularly useful for Windows where Python environments
are a mess and multiple different binaries are available
depending on how users install Python.

It does not solve the problem with virtualization and Python.

The core of Issues #771, #773, #957 is not directly addressed here
in the sense that a full Python environment manager has not
been implemented. Therefore, if `python3` is available the current
solution will not bother with `py` or `python`.

Fixes #1121
Fixes #957
Fixes #773
Fixes #771